[#48] Merge main & custom config, add MONITORING_FORM bool
